batch/embeddings/requirements.txt
openai>=1.0.0
httpx>=0.24.1
python-dotenv
